pyproject.toml
[tool.poetry]
name = "openclassrooms-projet-12"
version = "0.1.0"
description = ""
authors = ["devexplorer <contact@devexplorer.fr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
sqlalchemy = "^2.0.43"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"